+++ src/base/components/Loading.vue
@@ -2,7 +2,7 @@
   <div class="load_container" v-if="loading || loadingNum > 0">
     <div class="load_overlay"></div>
     <div class="load_image">
-      <!-- <img src="../assets/loading.gif" /> -->
+      <img src="../assets/loading.gif" />
     </div>
   </div>
 </template>