+++ src/i18n.js
@@ -1,23 +1,14 @@
-import Vue from "vue";
 import VueI18n from "vue-i18n";
 
-Vue.use(VueI18n);
-
 function loadLocaleMessages() {
-  const locales = require.context(
-    "../src/lang",
-    true,
-    /[A-Za-z0-9-_,\s]+\.json$/i
-  );
-  const messages = {};
-  locales.keys().forEach((key) => {
-    const matched = key.match(/([A-Za-z0-9-_]+)\./i);
-    if (matched && matched.length > 1) {
-      const locale = matched[1];
-      messages[locale] = locales(key);
-    }
-  });
-  return messages;
+
+  const en = require("../src/lang/en.json");
+  const fr = require("../src/lang/fr.json");
+
+  return {
+    "en": en,
+    "fr": fr
+  };
 }
 
 export default new VueI18n({