+++ src/lib-components/sections.vue
@@ -1,5 +1,4 @@
 <template>
-  <div id="app">
     <SectionsMain
       :admin="admin"
       :pageName="pageName"
@@ -7,12 +6,9 @@
       :lang="lang"
       @load="loaded"
     />
-  </div>
 </template>
 
 <script>
-import "bootstrap/dist/css/bootstrap.css";
-import "bootstrap-vue/dist/bootstrap-vue.css";
 import "../assets/scss/variables.scss"
 import SectionsMain from "../base/index.vue";
 export default {
@@ -33,11 +29,6 @@ export default {
   },
   components: {
     SectionsMain,
-  },
-  computed: {
-    variations() {
-      return [];
-    },
   },
   methods: {
     loaded(e) {